requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
apscheduler==3.10.4
pydantic==2.7.1
python-dotenv==1.0.1
google-generativeai==0.5.4
streamlit==1.35.0
requests==2.32.2